compat: cache Bleak version
The get_bleak_version call leads to opening a file to get the version.
This is fairly redundant as the Bleak version will be stable for the
lifetime of the process. Moreover it seems to trigger a small memory
leak according to tracemalloc:

/usr/lib64/python3.12/pathlib.py:1013: size=708 B (+59 B), count=12 (+1), average=59 B
/usr/lib64/python3.12/pathlib.py:1013: size=767 B (+59 B), count=13 (+1), average=59 B

So cache the value of the function to avoid a file access and a memory
leak.